requirements.txt
streamlit
requests
torch
openai-whisper
transformers
librosa
ffmpeg-python
yt-dlp
speechbrain
soundfile